Improve Ordering Gate behavior on out of order commits and proposals (#1243)
Previously ordering gate was not correctly handling proposals which would arrive before corresponding commits, and block height of the ledger was not used in deciding what to do with them. Also, there was a race condition on concurrent queue when someone was popping from the queue and another thread was accessing it, the proposal could be skipped by both threads.

Now we have Rx based concurrency, where each new commit or proposal event is handled by separate thread, which processes them in order, and block height is stored inside event. This means that we minimize number of shared variables between different events.

Signed-off-by: Nikita Alekseev <n.alekseev2104@gmail.com>