Add GetPendingTransactions query to shared model (#1532)
This PR adds an interface of GetPendingTransactions query to shared model.
Currently, the query does NOT perform actual information retrieving (will be implemented after appearing of transactions batches support in develop branch).
Additionally, this pr provides documentation for the new query.

It is a starting point for further development of GetPendingTransactions feature.

Signed-off-by: Igor Egorov <igor@soramitsu.co.jp>