Fix ToriiService Streaming Full Pipeline Test (#1615)
The test (StreamingFullPipelineTest) was failing on slow machines. 
Investigation showed that the root cause was the way thread scheduler works, so the code was generally valid.

Streaming client resubscription is a fully valid use case of streaming service.
The client can resubscribe to the stream if the previous stream was ended unexpectedly for the client.
Unexpectedly means that client requests the hash of a transaction that was definitely sent to Iroha and the stream reaches its end with one of the following statuses: `NOT_RECEIVED`, `STATELESS_VALIDATION_SUCCESS`, `STATEFUL_VALIDATION_SUCCESS`.

Before the fix test client was not able to resubscribe to the stream. Now it can.

```diff
cmake -H. -Bbuild
cd build
make torii_service_test
while test_bin/torii_service_test --gtest_filter=*ull* ; do :; done
+this loop should not be broken up by ASSERT_EQ(last_status, COMMITTED) error message
```

Signed-off-by: Igor Egorov <igor@soramitsu.co.jp>