Fix race bug in command service (#2167)
### Description of the Change
Fix bug when not committed/rejected transaction will not pass to Iroha's pipeline because it is a "replay".

### Benefits
Add fix for the problem. The idea of the fix to check only final statuses in the ram cache in command service. If there is a final status we will decline further checking of the transaction. Also, thanks @lebdron for cooperation in founding the bug.

Signed-off-by: Fedor Muratov <muratovfyodor@yandex.ru>